Refactor listing form into modular steps
