+++ client/src/__mocks__/fileMock.ts
@@ -1,0 +1,1 @@
+export default "test-file-stub";
